fix : working on modularizing code
